Python3 isoinstaller.
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/4325
Tested-by: gerrit-photon <photon-checkins@vmware.com>
