+++ common/data/packages_iso.json
@@ -1,9 +1,9 @@
 {   
     "packages": ["glibc","zlib","filesystem","file","gmp","libgcc","libstdc++","bash","sed",
-                  "bzip2","pkg-config","ncurses","ncurses-compat","cracklib","cracklib-dicts","cracklib-python",
+                  "bzip2","pkg-config","python3-curses","ncurses", "cracklib","cracklib-dicts","python3-cracklib",
                   "shadow","coreutils","grep","readline","findutils","xz","util-linux","e2fsprogs",
                   "libffi","expat","linux","cpio","Linux-PAM","attr","libcap", "systemd","dbus",
                   "gzip", "sqlite","nspr","nss","popt","lua","rpm","gptfdisk","tar", "librepo",
-                  "hawkey", "python2","python2-libs","pcre","glib", "tdnf", "python-requests",
+                  "hawkey", "python3","python3-libs","pcre","glib", "tdnf", "python3-requests",
                   "grub2-pc", "grub2-efi", "dracut", "curl", "ostree", "dosfstools"]
 }
