+++ installer/modules/m_updatehostname.py
@@ -12,6 +12,8 @@ def execute(name, config, root):
     hosts_file    = os.path.join(root, 'etc/hosts')
 
     with open(hostname_file, 'wb') as outfile:
-        outfile.write(hostname)
+        outfile.write(hostname.encode())
 
-    commons.replace_string_in_file(hosts_file, r'127\.0\.0\.1\s+localhost\s*\Z', '127.0.0.1\tlocalhost\n127.0.0.1\t' + hostname)
+    pattern = r'(127\.0\.0\.1)(\s+)(localhost)\s*\Z'
+    replace = r'\1\2\3\n\1\2' + hostname
+    commons.replace_string_in_file(hosts_file, pattern, replace)
